+++ mg_custom_nodes/KLL535_ComfyUI_Simple_Qwen3-VL-gguf_20260726/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Nightly
+- Added `📸 Load Video Fragment` node.
 - Added Bernini presets.
 - Add `_user_prompt_template` functionality. Now you can modify the user_prompt using a specified template (the name of which matches the system_preset)
 - Add `variables` input. You can now set any user placeholders in {} in the system and user prompts.